開き度合い
コントローラーの開き度合いを表示できるようにしました。
