+++ DemoGame002/Assets/Scripts/Ugoki_ConL.cs
@@ -7,8 +7,10 @@ public class Ugoki_ConL : MonoBehaviour
     public AudioSource audio;
     public AudioClip koukaon_1;
     public AudioClip koukaon_2;
-    public Text text;
-    int score;
+    public Text scoreText;
+    public static int score;
+    int score_L;
+    public AudioClip sound;
     // Start is called before the first frame update
     void Start()
     {
@@ -17,7 +19,7 @@ public class Ugoki_ConL : MonoBehaviour
 
     void Update()
     {
-
+        /*
         if (OVRInput.GetDown(OVRInput.Button.PrimaryIndexTrigger, OVRInput.Controller.LTouch))
         {
             OVRInput.SetControllerVibration(0f, 1f, OVRInput.Controller.LTouch);
@@ -27,8 +29,16 @@ public class Ugoki_ConL : MonoBehaviour
         {
             OVRInput.SetControllerVibration(0f, 0f, OVRInput.Controller.LTouch);
         }
-        text.text = "左コントローラ角度" + gameObject.transform.localEulerAngles.y.ToString();
+        */
 
+        //text.text = "左コントローラ角度" + gameObject.transform.localEulerAngles.y.ToString();
+
+        score = Ugoki_ConR.getScore() + score_L;
+        scoreText.text = "スコア" + score.ToString();
+        if (score > 10)
+        {
+            audio.PlayOneShot(sound);
+        }
     }
     void OnTriggerEnter(Collider col)
     {
@@ -36,13 +46,14 @@ public class Ugoki_ConL : MonoBehaviour
         {
             OVRInput.SetControllerVibration(0f, 1f, OVRInput.Controller.LTouch);
             audio.PlayOneShot(koukaon_1);
-            score = Ugoki_ConR.getScore()+2;
+            score_L += 2;
+             
         }
         if (col.CompareTag("bomb"))
         {
             OVRInput.SetControllerVibration(0f, 1f, OVRInput.Controller.LTouch);
             audio.PlayOneShot(koukaon_2);
-            score = Ugoki_ConR.getScore()-1;
+            score_L -= 1;
         }
     }
     void OnTriggerExit(Collider col)
@@ -52,5 +63,5 @@ public class Ugoki_ConL : MonoBehaviour
             OVRInput.SetControllerVibration(0f, 0f, OVRInput.Controller.LTouch);
         }
     }
-   
+
 }
