+++ DemoGame002/ProjectSettings/ProjectSettings.asset
@@ -128,7 +128,9 @@ PlayerSettings:
     16:9: 1
     Others: 1
   bundleVersion: 0.1
-  preloadedAssets: []
+  preloadedAssets:
+  - {fileID: 11400000, guid: 79849b69184304016ad965169d1d28f2, type: 2}
+  - {fileID: -4774611651135051186, guid: ed88faf1877f04f00873703ef21536d4, type: 2}
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
